fix: fix blueprintjs background color
The blueprintjs tree doesn't set a background, so the docs suggest
manually setting the background color. Since we don't have a custom
style sheet for this and we are using blueprintjs css from CDN, we
just hard-code the background color to `LIGHT_GRAY5` which is the
default background color.

This fixes the examples not being readable when docusaurus is in dark mode.